Created custom hook
